feat: restructure transaction controller
* fix: transform all ids to lower case
